8ball mode is now playable. However, there are is a very strange known bug: when the user plays first on the ball that is alone, the physics does completely strange things. I have no idea what actually the problem is. Because if you shoot on the triangle everything seems fine.
Another bug (probably the same cause) is that when the white ball falls into a hole , it does not collide with balls or walls anymore it just goes through. Again I can't figure out the cause.

some shots via banks are still incorrect (working on that one).

some balls get into each other on a collision ("verzahnt")

send me an email if you think the rules are not correct in the format: <what happened> <what you expected> <what the program did>